+++ internal/backend/externalclient/config.go
@@ -1,15 +1,16 @@
-package internalclient
+package externalclient
 
 import (
 	"go.autokitteh.dev/autokitteh/internal/backend/configset"
 )
 
 type Config struct {
-	InternalEndpoint string `koanf:"internal_endpoint"`
+	Enabled          bool   `koanf:"enabled"`
+	ExternalEndpoint string `koanf:"endpoint"`
 }
 
 var Configs = configset.Set[Config]{
 	Default: &Config{
-		InternalEndpoint: "http://localhost:9980",
+		Enabled: false,
 	},
 }
